+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* New back-end command, `match`, for non-prefix completion.
 * New user option `company-continue-commands`.
 * New back-end command, `annotation`, for text displayed inline in the popup
   that's not a part of completion candidate.